requirements.txt
deta
uvicorn
fastapi
air_telemetry
python-dotenv
requests